+++ src/components/Homepage/Landing.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 import Link from "next/link";
